Refactor timestamps para tener precisión de ms (uint64).

+++ ker/include/ker/cli_parser.h
@@ -11,7 +11,7 @@
 
 bool cli_parse_select(const cx_cli_cmd_t* _cmd, cx_err_t* _err, char** _outTableName, uint16_t* _outKey);
 
-bool cli_parse_insert(const cx_cli_cmd_t* _cmd, cx_err_t* _err, char** _outTableName, uint16_t* _outKey, char** _outValue, uint32_t* _outTimestamp);
+bool cli_parse_insert(const cx_cli_cmd_t* _cmd, cx_err_t* _err, char** _outTableName, uint16_t* _outKey, char** _outValue, uint64_t* _outTimestamp);
 
 bool cli_parse_create(const cx_cli_cmd_t* _cmd, cx_err_t* _err, char** _outTableName, uint8_t* _outConsistency, uint16_t* _outNumPartitions, uint32_t* _outCompactionInterval);
