+++ ker/include/ker/ker_protocol.h
@@ -72,7 +72,7 @@ uint32_t ker_pack_req_describe(char* _buffer, uint16_t _size, uint16_t _remoteId
 
 uint32_t ker_pack_req_select(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key);
 
-uint32_t ker_pack_req_insert(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key, const char* _value, uint32_t _timestamp);
+uint32_t ker_pack_req_insert(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key, const char* _value, uint64_t _timestamp);
 
 uint32_t ker_pack_req_journal(char* _buffer, uint16_t _size, uint16_t _remoteId);
