+++ lfs/include/lfs/lfs_protocol.h
@@ -55,6 +55,6 @@ uint32_t lfs_pack_req_describe(char* _buffer, uint16_t _size, uint16_t _remoteId
 
 uint32_t lfs_pack_req_select(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key);
 
-uint32_t lfs_pack_req_insert(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key, const char* _value, uint32_t _timestamp);
+uint32_t lfs_pack_req_insert(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key, const char* _value, uint64_t _timestamp);
 
 #endif // LFS_PROTOCOL_H_
