+++ mem/src/mem.c
@@ -506,7 +506,7 @@ static void handle_cli_command(const cx_cli_cmd_t* _cmd)
     char*       tableName = NULL;
     uint16_t    key = 0;
     char*       value = NULL;
-    uint32_t    timestamp = 0;
+    uint64_t    timestamp = 0;
     uint8_t     consistency = 0;
     uint16_t    numPartitions = 0;
     uint32_t    compactionInterval = 0;
